+++ src/search_graph/state_packer.cc
@@ -9,7 +9,7 @@ using std::vector;
 namespace pplanner {
 
 StatePacker::StatePacker(std::shared_ptr<const SASPlus> problem) {
-  size_t size = problem->n_variables();
+  int size = problem->n_variables();
   block_index_.resize(size);
   vector<int> bit_index(size);
   vector<int> bit_size(size);
@@ -17,7 +17,7 @@ StatePacker::StatePacker(std::shared_ptr<const SASPlus> problem) {
   int index = 0;
   int offset = 0;
 
-  for (size_t i=0; i<size; ++i) {
+  for (int i=0; i<size; ++i) {
     int n_bit = static_cast<int>(std::ceil(std::log2(problem->VarRange(i))));
     assert(n_bit <= 32);
     bit_size[i] = n_bit;
@@ -37,7 +37,7 @@ StatePacker::StatePacker(std::shared_ptr<const SASPlus> problem) {
   mask_.resize(size);
   shift_.resize(size);
 
-  for (size_t i=0; i<size; ++i) {
+  for (int i=0; i<size; ++i) {
     int right = 32 - bit_size[i];
     int left = right - bit_index[i];
     mask_[i] = UINT32_MAX >> right << left;
@@ -46,7 +46,7 @@ StatePacker::StatePacker(std::shared_ptr<const SASPlus> problem) {
 
   var_per_block_.resize(block_size_, 0);
 
-  for (size_t i=0; i<size; ++i)
+  for (int i=0; i<size; ++i)
     ++var_per_block_[block_index_[i]];
 }
 
@@ -66,7 +66,7 @@ void StatePacker::Pack(const vector<int> &state, uint32_t *packed) const {
 }
 
 void StatePacker::Unpack(const uint32_t *packed, vector<int> &state) const {
-  for (size_t i=0, n=state.size(); i<n; ++i) {
+  for (int i=0, n=state.size(); i<n; ++i) {
     int index = block_index_[i];
     state[i] = static_cast<int>((packed[index] & mask_[i]) >> shift_[i]);
   }
